+++ test/system_management_test.jl
@@ -12,19 +12,21 @@ using Test
 using DataFrames
 using Random
 using MathematicalSystems
+using JLD #need JLD in the module for reconstruction
 
 using AutomationLabsDepot
+
 
 @testset "List systems" begin
 
     # Create a project
-    project_name = "jean"
+    project_name = "cx"
     project_folder_create_db(project_name)
 
     # List the system project and test it is depicted
     list = list_system_local_folder_db(project_name)
 
-    @test_skip size(list) == (1, 6)
+    @test size(list) == (0, 6)
 
     # Remove the project
     remove_project_local_folder_db(project_name)
@@ -65,7 +67,11 @@ end
 
     # Load the system from the DataBase 
     sys_load = load_system_local_folder_db(project_name, system_name)
-    @test sys_load["system"] == sys
+    @test_skip sys_load["system"] == sys
+    @test sys_load["system"].A == sys.A
+    @test sys_load["system"].B == sys.B
+    @test sys_load["system"].X == sys.X
+    @test sys_load["system"].U == sys.U
 
     # Remove the system 
     remove_system_local_folder_db(project_name, system_name)